+++ OneBlink.SDK/models/FormPostSubmissionReceiptPdfConfiguration.cs
@@ -1,0 +1,16 @@
+using System;
+
+namespace OneBlink.SDK.Model
+{
+    public class FormPostSubmissionReceiptPdfConfiguration
+    {
+        public Guid? id
+        {
+            get; set;
+        }
+        public PDFConfiguration configuration
+        {
+            get; set;
+        }
+    }
+}